requirements.txt
ultralytics
opencv-python
paho-mqtt
